optimize structure
