+++ flink-table/flink-sql-parser/src/main/codegen/data/Parser.tdd
@@ -33,6 +33,7 @@
     "org.apache.flink.sql.parser.ddl.SqlAddJar"
     "org.apache.flink.sql.parser.ddl.SqlAddPartitions"
     "org.apache.flink.sql.parser.ddl.SqlAddPartitions.AlterTableAddPartitionContext"
+    "org.apache.flink.sql.parser.ddl.SqlAlterCatalog"
     "org.apache.flink.sql.parser.ddl.SqlAlterCatalogOptions"
     "org.apache.flink.sql.parser.ddl.SqlAlterDatabase"
     "org.apache.flink.sql.parser.ddl.SqlAlterFunction"
